tests/wpt/tests/navigation-api/updateCurrentEntry-method/same-document-away-and-back-location-api.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(async () => {
    let start_length = navigation.entries().length;
    let entry0 = navigation.currentEntry;
    let start_index = entry0.index;

    let navState = { statevar: "state" };
    navigation.updateCurrentEntry({ state: navState });
    assert_equals(navigation.currentEntry, entry0);

    location.href = "#2";
    let entry1 = navigation.currentEntry;

    assert_equals(navigation.entries().length, start_length + 1);
    assert_equals(entry0, navigation.entries()[start_index]);
    assert_equals(entry1, navigation.entries()[start_index + 1]);

    assert_equals(entry0.getState().statevar, "state");
    assert_not_equals(entry0.getState(), navState);

    assert_equals(entry1.getState().statevar, "state");
    assert_not_equals(entry1.getState(), navState);
    assert_not_equals(entry1.getState(), entry0.getState());

    history.back();
    window.onpopstate = t.step_func_done(() => {
      assert_equals(navigation.entries().length, start_length + 1);
      let back_entry = navigation.currentEntry;
      assert_equals(back_entry, entry0);
      let back_state = back_entry.getState();
      assert_not_equals(back_state, navState);
      assert_equals(back_state.statevar, "state");
    });
  }, 0);
}, "entry.getState() behavior after navigating away using the location API, then back");
</script>
