LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/same-document-away-and-back-location-api-expected.txt


FAIL entry.getState() behavior after navigating away using the location API, then back assert_equals: expected 3 but got 4
